Move the database to SQLite on the persistent disk
Liara's managed Postgres turned out to cost about 550,000 Tomans a month
once the mandatory feature bundle is added to the plan price — the plan
figure alone is only half the bill. For a site whose entire dataset is
about 150KB, with far more reads than writes and a single app instance,
that is not a good trade.

SQLite on the mounted disk costs nothing extra and is faster here: queries
run in-process with no network round trip. WAL mode keeps a save in the
panel from blocking visitors, and foreign keys are enabled per connection
since SQLite leaves them off by default.

Nothing above the database layer changed. The repos, services, cache,
media pipeline, admin panel and every page work exactly as before, which
is the point of having put them behind that boundary. The dialect-specific
SQL that did need changing was small and local: ilike became like,
greatest became SQLite's multi-argument max, and the count casts went
away.

Development and production now run the identical engine rather than PGlite
standing in for Postgres, so a query that works locally cannot behave
differently on the server.

The tradeoff is backups, which are no longer someone else's job. There is
a db:backup script using VACUUM INTO, so it produces a consistent copy
while the site is serving rather than a possibly-torn file copy, and
DEPLOY.md now has a section on running and storing them.

One disk named data holds both the database and the uploaded images, and
tsx moved to dependencies so the seed, admin and backup scripts can run
inside the container.

+++ src/server/services/outbox.ts
@@ -120,15 +120,15 @@ export async function outboxStatus(): Promise<OutboxStatus> {
 
   const [[pending], [failed], [sentToday]] = await Promise.all([
     db
-      .select({ value: sql<number>`count(*)::int` })
+      .select({ value: sql<number>`count(*)` })
       .from(outbox)
       .where(and(isNull(outbox.sentAt), sql`${outbox.attempts} < ${MAX_ATTEMPTS}`)),
     db
-      .select({ value: sql<number>`count(*)::int` })
+      .select({ value: sql<number>`count(*)` })
       .from(outbox)
       .where(and(isNull(outbox.sentAt), sql`${outbox.attempts} >= ${MAX_ATTEMPTS}`)),
     db
-      .select({ value: sql<number>`count(*)::int` })
+      .select({ value: sql<number>`count(*)` })
       .from(outbox)
       .where(sql`${outbox.sentAt} >= ${startOfDay}`),
   ]);
